refactor blog module to news models and background task processors
