+++ docs/featuremap/types.csv
@@ -14,6 +14,7 @@ Fragment
 Library
 Model
 Process
+Query
 Schema
 Script
 SourceData
