+++ portality/models/__init__.py
@@ -21,6 +21,7 @@ from portality.models.cache import Cache
 from portality.models.openurl import OpenURLRequest
 from portality.models.provenance import Provenance
 from portality.models.background import BackgroundJob
+from portality.models.news import News
 
 import sys
